1.修改智能闹钟界面 2.修正时间早于8点的日期显示错误的bug 3.修改更新固件的流程（当升级失败断开后，重新连接，发送复位命令，手表就会退出升级模式） 4.修改登陆界面 5.提交到appStore和fir上，版本1.1.5

+++ WMSPlusdot/DFUOperations.m
@@ -339,7 +339,7 @@ NSDate *startTime, *finishTime;
 {
     finishTime = [NSDate date];
     self.uploadTimeInSeconds = [finishTime timeIntervalSinceDate:startTime];
-    NSLog(@"upload time in sec: %u",self.uploadTimeInSeconds);
+    NSLog(@"upload time in sec: %lu",(unsigned long)self.uploadTimeInSeconds);
 }
 
 #pragma mark - BLEOperations delegates
@@ -388,7 +388,7 @@ NSDate *startTime, *finishTime;
 
 -(void)onFileOpened:(NSUInteger)fileSizeOfBin
 {
-    NSLog(@"onFileOpened file size: %d",fileSizeOfBin);
+    NSLog(@"onFileOpened file size: %lu",(unsigned long)fileSizeOfBin);
     binFileSize += fileSizeOfBin;
 }
